clients: drop run_with_tb
It is subsumed by tmp_tigerbeetle, which is a bit more flexible and
doesn't rely on the shell.

The last usage of run_with_tb was for client's readme validation, but
we've replaced that with samples validation.

As a corollary, drop most of the over remnants of old validation code.