java: don't use checked exceptions for our own programming errors
RequestException is thrown when the underlying C library detects that
the request is invalid --- the op or the size of message body is wrong.

However, the public API of of our Java client ensures that these are
always correct: it is impossible to trigger these errors unless:

* there's a bug in our C or Java code
* there's a version mismatch between our Java and C code (which is
  unlikely, as they come from the same jar)

So, there's absolutely no reason to tax the caller with declaring and
handling this exception.