VSR/CheckpointTrailer: Encode/decode using grid blocks
Instead of encoding/decoding the trailer into/from a `buffer`, encode/decode it directly to/from grid blocks.

Benefits:
- One less copy (of all trailer data) during every checkpoint.
- Once we have a grid pool, blocks can be acquired lazily, and when not used (which is most of the time, since the free set is _usually_ very small) they contribute to the grid cache.

To keep implementation simple (avoiding modifications to `ClientSessions.{encode,decode}`), one assumption this makes is that the `ClientSessions` trailer fits in a single grid block. (The ClientSessions trailer is fixed-size.) This was already true for all of our configurations.