VSR/Sync: Panic on checkpoint divergence
In the original state sync design, checkpoint divergence (e.g. due to storage engine nondeterminism) was tolerated – within a given bound, the cluster could use state sync to automatically recover. The rationale for automatically recovering from checkpoint divergence was that it would simplify replica version upgrades. In particular, it was to add some guardrails case an upgrade that was not expected to change determinism did change it.

However, after some discussion we decided that storage engine nondeterminism is Not Allowed. (Specifically, mixed-version clusters will be forbidden.) If divergence is detected, we will now abort, to escalate to manual operator intervention.