VSR: Assume all checkpoints are canonical
Assume that all checkpoints are canonical checkpoints.

In the original design of state sync, the distinction of canonical/uncanonical checkpoints enabled the cluster to (automatically) recover from a divergent checkpoint.
Since checkpoint divergence is now forbidden (on penalty of process-abort) this distinction is no longer necessary.

This fixes the liveness issue formerly guarded by `Simulator.core_missing_checkpoint()`, and avoids a crash in that same function:

- Fixes https://github.com/tigerbeetle/tigerbeetle/issues/1382
- Fixes https://github.com/tigerbeetle/tigerbeetle/issues/1430

The new logic to answer "Can replica X sync to checkpoint Y?" is:
1. the new checkpoint must advance the replica's op, and
2. the new checkpoint must either be committed atop _or_ more than one checkpoint ahead of our current checkpoint.
    - If a checkpoint is committed atop, that guarantees that a (commit) quorum of replicas has reached that checkpoint.
    - See `Replica.jump_sync_target()` for more details about syncing to an idle cluster.