workbench: initial
 ## What & Why

Add a workbench --- a simple HTML page, maintained as a part of
TigerBeetle monorepo, whose purpose is to serve any kind of developer
related information in a low-effort way. Basically, this is GUI for all
our automation.

Currently, workbench displays two pieces of information:

1. Release rotation (subsuming the python script)
2. Results of macro benchmarks across commits (binary size & build time)

To enable 2, this also adds some continuous benchmarking (CB)
infrastructure.

 ## How: Benchmark

CB is triggered by an GH action on the main branch. This action
intentionally isn't a CI gate (yet?). The idea is that we want to track
long-term trends in numbers, and find obvious regressions within a week.

CB is orechstrated by the `scripts/workbench.zig` script. There's no
specific rules what goes there, but, for starters, track just two
numbers:

- the size of a release binary,
- the time to build the software.

These probably aren't the most impactful things to track, but the idea
here is to establish a process of tracking. Adding _more_ things should
be relatively trivial, once the base infrastructure is in place.

Benchmark results across many runs are stored as a JSON file in the
<https://github.com/tigerbeetle/workbenchdb> repository, using `git` as
a distributed database of sorts.

Specifically:

- `workbench/data.json` is the file (the directory to allow "log
  rotation" once a year)
- it is a newline delimited JSON --- each line is a JSON object
- each objects records a "run" --- a set of measurements at a particular
  commit and a timestamp
- a measurement has a label, a value, and a unit (bytes, ms, etc)
- aggregating measurements with the same label across runs gives a time
  series
- the set of measurements can change over time --- different runs might
  have different sets of measurements
- unit is convenient both for humans and for the front-end to display
  results without hard-coding a list of measurements.
- as a TODO, it would perhaps be nice to track hosts cpu, memory and an
  OS in the run.

It is assumed that a single workbench is run at a time, so no provision
for merge conflicts is made.

  ## How: Workbench

A GitHub pages without any build step --- just some static files in
`src/workbench` directory. If this grows big, we'll probably want to
switch to TypeScript at least, but maybe it needn't grow big.

To display benchmark results, the data is `fetch`ed from
`raw.githubusercontent` and displayed using <https://apexcharts.com>. No
affiliation with this particular library, it's just something I saw on
<https://deno.com/benchmarks> and I quite like that page.