ci: drop lint_tigerstyle
lint_tigerstyle has some great aspirations, but it clearly isn't usable
and isn't used in its current form --- it prints a couple of hundred
lines to stdout, and a couple of hundred lines to stderr, and no one is
reading those lines.

It computes two things:

1. assertion density (number of asserts per function)
2. function length

One feels like a useful statistics, but it perhaps should be exposed as
a part of our overall metrics package, alongside with binary size and
TPS.

The second one is more contentious (we clearly have quite a few quiet
long function!), and it only makes sense to tackle that after we get
line length in order (for which we have a ratchet already, but still
need to go through the naughty list of exceptions).