vsr: isolated replica can't lock out a standby from a cluster
Standbys don't participate in view changes, so switching to
`.view_change` is useless. This also prevents an isolated replica from
locking a standby into a view higher than that of the rest of the
cluster:

- R0, R1, R2, S0 are in view=0
- connectivity temporarily degrades, such that R1 and R2 send SVCs
- R1 gets a quorum of SVC
- connectivity between R0 and R2 is restored, but R1 gets isolated from
  both of them.
- At this point:
    - R0, R2 is a functioning cluster in view=0
    - R1 is an isolated replica view=1

If R1 can talk to S0, it is important that S0 _doesn't_ advance its
view, otherwise it'll get isolated from R0 and R2, which remail fully
available.

Closes: #1455
SEED: 8961816063059206802