vsr: more logging around advancing commit_max
This line is pretty crucial --- this is _the_ moment when we say "yep,
this prepare here is committed and a part of the official history", if
we are a primary.

Looking at this more generally, I wonder if perhaps we should
restructure the control flow here a bit, so that `commit_max` is
advanced _before_ we start committing, so that we don't mix up control
plane (deciding what's committed) and data plane (actually committing
stuff).

But just logging this seems like a good start.