vsr: liveness when SV strands checkpoint boundary in an idle cluster
An SV might include messages from the current and the next checkpoint.

Installing headers from the next checkpoint would be unsafe, so replica
effectively accepts only _part_ of an SV message if it hasn't
checkpointed yet.

What makes this tricky is that "skipped" headers are from the previous
view.

So, if a a replica sees such a header/prepare after it checkpoints, it
will _still_ ignore it: this header is from the previous view, so it
could have been truncated.

For prepares from older views, a replica needs primary to positively
tell it that, yes, that prepare is part of the canonical history.

To fix this, detect if our view headers are ahead of our op during
repair, and request a start view in that case.

SEED: 15556801325249033142