vsr: prevent crash due to inconsistency between client_sessions and client_replies
During state sync, we don't bother to "cancel" in-flight client_replies.
This is OK: we mark entries as faulty during state sync, and in-flight
outdated writes can't clobber that state.

This means though that any in-flight write might not actually match the
contents of client_sessions table, tripping an assert in
`read_reply_sync`.

To correct this:

- explicitly check for matching checksum, return `null` if we are
  writing a wrong reply.
- during state sync, set faulty client entries right when we decode
  client_sessions, to make sure clinet_sessions and client_replies
  remain consistent.

An alternative fix here would be to add cancellation to
`client_replies`, but:

- That seems like it would increase the state space and in general be
  more complicated.
- The current code already anticipates concurrent state syncs (see the
  comment in `write_reply` about updating `faulty` at the start, rather
  than at the end, of writing process).

SEED: 15814337035080855944
Closes: #1486