dotnet: use random data file for testing
On windows on CI, this at times throws AccessDenied. I haven't got to
the bottom of it, but my experience tells me that one does not simply
remove a file on windows:

https://github.com/build2/libbutl/blob/d249dff880edd1520a979279934afc0a6370bd67/libbutl/filesystem.cxx#L1784

So, my hypothesis here is that, because all tests try to re-use the same
file, we are hitting some standard windows gotcha.

Let's do it properly by generating a random file instead.

Note: integration tests using tmp_tigerbeetle already make use of a
temporary directory, so the equivalent code executes as a part of Zig's
stdlib.