cli: introduce tigerbeetle benchmark
`tigerbeetle benchmark` runs a standard benchmark on the current
hardware. It is built into the production binary, and is intended to be
used "in the field".

At the same time, this an internal command --- we won't give any
stability guarantees with respect to its output format. Similarly, we'll
adjust benchmarking options liberally as we go. The options aren't
documented yet for this reason.

The benchmark runs in two "modes":

- if the `--adresses` argument is passed, than the benchmark issues load
  against the specified cluster.
- without `--adresses` argument, the benchmark spins up a temporary
  single-node cluster on the current machine.