vsr: consider state sync done only when all replies hit the disk
Before `sync_content_done` could return true while some repaired replies
are still in flight. This is a durability issue: if replica completes
the sync, but then crashes _before_ the reply is fully written, it won't
attempt to fix it subsequently.

And, naturally, this will be caught by the VOPR, which notices that data
files are different at a checkpoint between replicas.

SEED: 11435969822295410219
Closes: #1494