[Experimental] Implements account history as two grooves
The simplest (and fastest) way to insert, it requires no changes to the Groove, but it is complex to query:

Two grooves "debits_history" and "credits_history" with the same Object containing the four balances + the transfer timestamp.

No indexes for fast ingestion.

We need two Grooves because the Transfer timestamp is the PK and it needs to be unique on each Groove.

We can retrieve by using the existing indexes on debit_account_id and credit_account_id that yield the transfer timestamp needed for the lookup.

The tricky part is that we must support multi-Groove queries to join two indexes, each one performing a lookup into a different groove!