Benchmark: Forward replica process stderr
Forward the replica child process' stderr to the parent process.

(Sometimes I like to add timers to specific parts of the code before running the full benchmark – this way those results are easily available.)

Also if there were any non-fatal issues with the replica, this would help to surface them.