vsr: don't lose SV headres during view change
This commit is a stepping stone to fixing #1510.

It deals with the following effect: a replica receiving an SV might not
install all headers from the SV. Specifically, headers from the next log
wrap are not added to the journal. As a result, such headers don't
necessary survive into a DVC for a subsequent view change.

We special case (via `start_view_into_do_view_change`) scenario where
the head op gets lost this way. However, ops other than the head might
get lost as well:

- prepare_max=31
- replica receives SV with ops 31..32
- replica does not install op=32
- replica receives a prepare with op=33
- during view change, head op=33 is larger that view headers op=32, so
  SV_into_DVC logic doesn't get triggered.
- as a result, op=32 is absent from the DVC

This commit makes SV_into_DVC logic more precise: rather than
constructing DVC from SV or from journal, the DVC is constructed from
the union of both.

Right now, the only effect of this change should be that more headers
are present in DVC, and, as a result, a would-be primary should require
less repair.

However, this change also establishes a new, powerful invariant:

A DVC with log_view=v is now guaranteed to include all headers from
previous views, that survived into view=v. That is, if a primary decides
that a particular header should not be truncated, that fact is
guaranteed to be present on _any_ replica which starts view=v. This in
turn will allow us to strengthen the nack protocol!

Consider the following DVC quorum:

dvc replica=2 log_view=2
    op=12 view=2
    -
    -

dvc replica=1 log_view=1
    op=12 view=1
    op=11 view=1
    op=10 view=1

Here, `op=10` can be truncated --- although we don't the canonical
header for this op, we know that it can't be op=10 view=1. Because
view=1 < log_view=2, if the header is canonical, it must be a part of
replica=2's dvc.