devhub: switch to a line graph and other misc enhancements
I'm not sure if there are some ardent bar graph supporters here, but I think
it's easier to see the change over time on a line graph... The y-axis range
becomes relative now too.

Additionally, open tooltips to the commit rather than the range and show the
date on the x-axis as it has more meaning than the commit sha (still as a
category, though, to keep the points equally spaced).

Switch the date formatting to use .toLocaleString() for a nicer formatted
date (/ time).