vopr: false negative when core has a header break
The previous logic looked for the first uncommitted op (on any replica)
and checked whether the cluster's core misses that op.

However, it might be case that, even though the earliest uncommitted op
is present, it still can't be committed due to a header break down the
line:

    commit_min           break      op

       47      48         67        97

With the above picture, if `op=67` is missing from the cluster, then
`op=48` won't be committed, even if the replica has it locally. Before
committing `op=48`, the replica needs to verify that it hash-chains to
`op=97`, which it can't do due to an (unrepairable) break.

To correct this, explicitly look for the earliest op between commit_min
and HEAD which is absent on the replica.

Closes: #1564
Seed: 5630435061244550066