vsr: don't repair potentially unavailable ops
op_repair_min tries to be smart and optimistic and repair _more_, which
sometimes lead to unavailability, due to two issues:

- primary tries to repair more than backups, but because backups repair
  less, those prepares might be unavailable! That is, we might suffer
  _more_ than `f` failures  for particular prepare over time, because it
  is not continuously repaired on the backups
- after state sync, replicas simply do not have any prepares beyond
  checkpoint.

Simplify op_repair_min logic to always repair from the latest
"confirmed" (committed on top of) checkpoint.

That is, replicas _right_ at the trigger repair the entire log wrap, in
addition to having the checkpoint. This ensures smooth hand off of
durability: prepares stop being repaired only when there's a quorum of
replicas at a checkpoint including the prepare.