devhub: duplicate benchmark data to nyrkio
https://nyrk.io is a cloud CI benchmark tracking service.

It is very similar to what devhub is doing, except that it actually is a
real project, and not just me copy-pasting JavaScript from
https://deno.com/benchmarks.

Otherwise, it is very similar --- their data model is isomorphic to what
we already use.

In this PR, we duplicate data to both devhubdb and nyrk. I think we'll
continue in this mode for some time, but I can see us switching to nyrk
100% in the future.

In a follow up, I do want to switch to nyrk-shaped json for devhubdb as
well, as I like their version more (esp grouping attributes _separately_
from the timestamp, which is the primary key), but I don't want to
change stuff on our end right away.

As nyrk is a 3rd party service, I don't fail CI job if we fail to
upload.