tests: terminology change covered -> marks
After chatting today with Joran I realized that "coverage" is the wrong
word here. This is a very poor implementation of coverage indeed, as it
requires you to manually annotate the code, instead of relying on the
computer to do the job.

But coverage isn't the point. What we actually do here is traceability
--- linking two artifacts, source code and tests, together, to make sure
that they motivate each over.

Coverage answers the question "Is this tested?". Marks answer "Why this
code needs to exist?"