vsr: add entire header to checkpoint state
To function, a replica needs a checkpoint and a header in WAL. A header
is required to make "head op is in journal" invariant work.

Currently after state sync a replica has only the checkpoint, with the
headers absent. We paper over this by transitioning the replica to
`.recovering_head`, but that creates availability issues, as in

    test "Cluster: sync: view-change with lagging replica in recovering_head"

To make the cluster more available, include an entire header with the
checkpoint, and not just op/checksum pair.