vopr: default to short log
It seems the original logic here was "when we debug a seed, we run in
debug, and we want full log for a single seed", but this doesn't work
well in practice, because full log is much too verbose. If I run with
log-full, I pretty much always want `-Drelease` and also to pipe the log
to a file.

Let's simplify this.