Test/ReplySequence: Set up ReplySequence to allow skipped prepares
The `Simulator`'s `ReplySequence` assembles replies (from the cluster to all clients) into commit order.
(Replies may arrive out of order, but the workload checker needs to process them in commit order so that the workload can correctly reproduce the cluster's sequence of events.)

Right now `ReplySequence` has a comprehensive view of replies: currently all `command=request`s originate from a client (`client_id ≠ 0`).

But shortly we will introduce request types that originate from the primary itself (`client = 0`):

- upgrade operations
- expiry messages, to drive transfer timeouts.

As such, this commit updates `ReplySequence` so that we will be able to skip past upgrade/expiry operations that the clients will never receive a reply for. This will work by changing the `else` branch of the `on_cluster_reply` loop to:

    if (prepare_header.client == 0) {
        simulator.reply_op_next += 1;
    } else {
        break;
    }