VSR: Add Replica.releases_bundled
This is presently stubbed out.
- For real (multiversion) binaries, this will be set to a list of releases compiled into the binary.
- For testing, we will manually set it to a list of releases (which may change between restarts).

`vsr_releases_max` is a new config/constant which defines the maximum number of releases that can be advertised by a replica in a `ping` message. (So this is effectively the maximum number of releases that can be compiled into a multiversion binary).
I set it to `64` right now, which I chose somewhat arbitrarily: We release on a weekly cadence, there are 52 weeks/year, rounded up to the nearest power-of-2 is 64.

Depending on if (and how well) we can compress multiple releases into a single binary, it might be too low or too high.