VSR: Ping bundled releases
To determine when the cluster can upgrade, each replica advertises which releases it is able/willing to upgrade to.

Note that pings will include all bundled releases -- newer, older, and currently-running. For the purposes of upgrades, only the newer versions actually matter, but the list isn't that large, so there isn't really a benefit to discarding/filtering the information out.