VSR: Downgrade during Replica.open()
When a replica first starts (`./tigerbeetle start`), the multiversion binary will execute the highest available version.

If we open the superblock and see that our code release is higher than the superblock `CheckpointState`'s release, then we will `release_execute()` into the superblock's release.
(If that release is unavailable, that is an operator error – we will panic.)