cli: add explicit `--standby` argument
Before:

   * `--replica=1 --replica_count=2` creates an active replica with index=1
   * `--replica=2 --replica_count=2` creates a standby with index=2

Now:

   * `--replica=1 --replica_count=2` creates an active replica with index=1
   * `--standby=2 --replica_count=2` creates a standby with index=2
   * `--replica=2 --replica_count=2` is an error

That is, the semantics of the argument is the same, its replica index
stored in the superblock, but the name must be `--replica` for active
replicas and `--standby` for standbys.

This binds intention into the CLI and makes it more self explanatory and
less error prone.

Additionally, we have an easier time "hiding" `--standby` feature until
we need it for reconfig, backups, cdc, or canaries.