cli: introduce concept of experimental features
Some flags are experimental: intentionally undocumented and are not a part of the official
surface area. Even experimental features must adhere to the same strict standard of safety,
but they come without any performance or usability guarantees.

Experimental features are not gated by comptime option for safety: it is much easier to review
code for correctness when it is initially added to the main branch, rather when a comptime flag
is lifted.