VSR: Send command=commit before checkpoint begins
Send a commit message immediately before a primary begins to checkpoint, rather than waiting for our commit timer.

This is useful when this checkpoint is an upgrade, since we will need to restart into the new version. We want all the replicas to restart in parallel (as much possible) rather than in sequence (to minimize downtime due to recovery/replay).