vsr: fix liveness issue when recovering head replica can't state sync
To recap, the issue:

* we can't persist checkpoint until our view is up to date, otherwise we
  observe a header break between the log and and checkpoint
* so we wait until our view_durable is up-to-date
* but that doesn't work for recovering head replicas, as it doesn't update its view durable
* and we _can't_ just excempt recevering head replica from this logic,
  as it might crash and restart into .normal due to transient failure

The new insight is that we can actually fix the latter, and make
`.recovering_head` case sticky across restarts! If a replica reastarts
and notices that it's checkpoint.header view is greater than view/view
durable, that necessary means that it crashed in recovering head state,
and needs to continue to recover its head!