vopr: state checker knows about all commits in liveness mode
We currently switch to liveness mode only when everything is committed
by the cluster, which is not as live as we ideally want to get, but
that's a start.

In particular, when liveness mode checks for too many corruptions, it
makes use of the fact that for every op the state checker knows the
correct header.

Before upgrades, this invariant was enforced by checking that every
request received a reply. But upgrade requests don't generate replies!

So it might be the case that, at the time we transition to liveness
mode, there's still an uncommitted op!

Fix this by checking that at least one replica upgraded all the way up
to the latest release.

Seed: 2749715711070932366
Closes: #1697