Config: Inject release version via CLI
Currently this uses `-Dversion`... sort of. It is injected as a separate build option, b/c unlike the preexisting `version`, we don't want the release version to fall back to the git tag (which is not necessarily a valid version). Instead we want it to default to the "dev" release (`0.0.1`).

This does seem fragile though.