vopr: upgrades cancel async work properly
Normally, executing upgrade is a noreturn function. In the simulator, it
is not, and instead the replica is restarted on the next tick.

The problem with this is that there might be some callbacks scheduled on
the storage, which fire before the actual upgrade happens, and observe
an inconsistent state (as, aftre transitioning to a new release replica
doesn't bother to repair local invariants).

Cancel the storage immediately upon scheduling upgrades to prevent those
callbacks from firing.

Seed: 2843879175163466061