vsr: commit_dispatch doesn't cancel cleanup
commit_dispatch starts off as

    fn commit_dispatch(self: *Self, stage_new: CommitStage) void {
      assert(self.commit_min <= self.commit_max);
      assert(self.commit_min <= self.op);
      if (self.syncing == .canceling_commit) {
          return self.sync_cancel_commit_callback();
      }

That is, every commit stage is a cancellation point. Notably, this
includes the `.cleanup` stage, which:

* releases the commit_prepare
* transitions to the next release

Not rleasing prepare is ok --- sync_dispatch would handle that in
cancelling grid stage

But not doing release transition is bad, as we rely on our release and
superblock release being in sync all the time.

To fix this, just make the cleanup stage of commit flow sync.

Seed: 14169780878022532988