bus: alignment of header is not assumed
This `@alignCast` could fire. There are several options to fix it:

1. Require that message size is a multiple of header's alignment.
2. At the transport layer, artificially pad messages to alignment.
3. Copy the data to fix the alignment.

I think 3 makes most sense:

1 doesn't really work, as state_machine replies are not aligned. A
replica is a sequence of (u32 index, u32 error code) tuples, and they
are 8-byte and not 16-byte aligned.

2 feels hacky in general and changes our network protocol.
Additionally, there could be all kinds of encryption&compression
layers in the transport, and it feels odd to try to maintain physical
alignment through all that.

3 initially felt wrong to me, but I've grown to like it. The thing is,
we _already_ have a memcpy here. If a replica gets two messages in a
single `.read` call, the second one would be copied over to a new
message anyway.

Although the current implementation copies the entire header for
simplicity, we actually don't need to do that. Message bus uses
header to calculate checksum and to look at a couple of fields
(cluster, client, replica, version). It should be perfectly fine to
do all that on an unaligned data.

Closes: #1518