vsr: assert that SV doesn't go past op_repiar_min
We require that the primary always includes as much headers as fit into
an SV. That is, the header can be omitted _only_ if we get all the way
to op=0.

But primary doesn't have literally _all_ messages, its log cuts off at
op_repair_min.

This all works out: if primary is at a checkpoint, it has at least
lsm_batch_multiple headers on top of it, and that's larger than
`view_change_headers_suffix_max`. But we should assert this!