compaction: use FIFO as a block_pool instead of trying to slice arrays
Previously, we tried to slice up an array of compaction blocks and hand them
out. Due to an error in the calculation of block_count_bar, not enough blocks
were reserved, so beat scoped blocks could alias higher bar scoped blocks.

This was caught by asserts, when it did eventually happen, but rather than
just fixing the math let's make it much harder for the situation to arise.

Instead of an array, blocks are now popped off a simple FIFO, and pushed
back on when they're done. Additional asserts have been added, ensuring
values match what we expect, and all the setting of state
(eg, `block.stage = .free`, `block.next = null`) have been removed and is
instead replaced with a proper block lifecycle.