docs: DEEP_DIVE & demos are retired in favor of quick start and repl
DEEP_DIVE isn't _very_ deep actually --- it just shows how to talk to
tigerbeetle. But we already have that flow documented in a quickstart
using the repl.

Similarly, demo code isn't really looked at by anyone, we don't run it
on CI and don't really know whether it works or not.

When we do a supported Zig client, we should have samples that mirror
the ones for all other clients, but, for know, its better to remove it
to simplify  the build and to reduce the amount of untested code in the
repository.